small fixes
git-svn-id: https://workrave.svn.sourceforge.net/svnroot/workrave/workrave/workrave/trunk@30 2b0e655b-6f86-497a-824e-95ef0e1e0d90

+++ src/ActivityStateMonitor.cc
@@ -109,7 +109,7 @@ ActivityStateMonitor::mouse_notify(int x, int y, int wheel_delta)
       int delta_x = x - prev_x;
       int delta_y = y - prev_y;
       
-      total_movement += sqrt(delta_x * delta_x + delta_y * delta_y);
+      total_movement += int(sqrt(delta_x * delta_x + delta_y * delta_y));
       
       prev_x = x;
       prev_y = y;
@@ -128,7 +128,7 @@ ActivityStateMonitor::button_notify(int button_mask)
       int delta_x = click_x - prev_x;
       int delta_y = click_y - prev_y;
      
-      total_click_movement += sqrt(delta_x * delta_x + delta_y * delta_y);
+      total_click_movement += int(sqrt(delta_x * delta_x + delta_y * delta_y));
     }
   click_x = prev_x;
   click_y = prev_y;
